Flows not deleted resulting in vrf del pending.
Problem:
Flows were expecting index to be allocated by vrouter, however vrouter sends
error for same. On receiving this event, these flows are marked as short flow.
Stats collector visits these flow and enqueues delete for these flows. Delete
routine checks if index is allocated. If its not then deletion is skipped
assuming that vrouter still has not added index. It does not check that index
generation resulted in error and now flow has to be deleted.

Solution:
In delete check if flow has been marked as short flow because of index
allocation failure and let it get deleted.

Change-Id: I71e8c3b6ddec7fb3884cc613c9f79d2fdb15301f
Closes-bug: 1531295