GracefulRestart Phase 2 -- Handle routing-instance deletions during GR
When GR is in progress, all routes and subscription states (in BgpXmppChannel)
are retained in the hope that agent resubscribes after GR. But if the config
changes between the time agent went down and came back up, new agent may not
re-subscribe to all the instances it had subscribed before. Such sticky
subscriptions must also be cleaned up properly using mark and sweep approach
(Similar to how routes are handled)

Mark and sweep approach is already implemented for routes in PeerCloseManager.
BgpXmppChannel rides on this to get necessary callback for

o When to mark all subscription states as stale
o If agent re-subscribes after restart, stale flag is cleared in the
  subscription state.
o When to sweep all subscription states (and delete still stale entries)

Only after all (still) stale subscription states are deleted, routing-instance
deletion process can resume and complete.

Btw, during GR, all route-targets retained as is, similar to how routes are
retained. At the moment, the rtarget entries are not individually marked
stale (and then swept). Instead, it is handled by marking the subscription
states which is already maintained on a per instance basis in BgpXmppChannel

Added test cases to cover many (but not all) scenarios

Change-Id: Id8225c5ee3de2efbc56c3fc0d0a613c7bb8e45c8
Partial-Bug: #1537933