Remember the order in which elements are added in dependency tracker
The problem is happening because we are trying to create VRF assign
rules even before service chain has been created. This can be
avoided if dependency tracker remembers the order in which it walked
the resources. Hence, changing the datatype to OrderedDict instead
of the regular dict.

Also, added a check for ip version before adding vrf assign rule
so that prefix length can be added correctly

Change-Id: Ic38b0b380fa70bedb3e357dfe3bfff6104ccae27
Closes-Bug: 1539293