Create the service chain with least one vm or port tuple
Currently, if any of the vm or port tuples of a service instance is not in good
state, we don't create the service chain. However, even with one of the VMs up,
the service chain can function correctly. With this chain, service chain
will only go to error if none of the VMs/port tuples are usable.

Change-Id: Ieddafb56e0b6be3a093111c10f2bdde8f486b181
Closes-Bug: 1541940