Closing the pipe FD while deleting the instance task
When instance manager spawns a task it opens a pipe between parent
process and child process. The child process stderr is linked to write
end and Agent reads the error messages of child process from the read
end. Data from this pipe is read using asio and if any error is found
while reading the pipe fd is closed. Along with this pipe event,
SIGCHILD is also registred to track the child process exit. As part of
SIGCHILD instance task is deleted. If pipe event is received later than
SIGCHILD event, dereferencing the instance task as part of pipe event is
leading to crash.

As a fix, the pipe descrptor is closed if open in the destructor of
instance task.

Change-Id: I4d458d65d55803142ea3ba5a687b16b24b4ed859
closes-bug: #1542159