Instance ip delete only on last vmi delete
Fix a bug wherein the instane-ip gets deleted when
service-instances are scaled down or an error service
vm is deleted. Instead of cleaning up the instance ip
the fix is to remove the link to the vmi and only when
the last vmi has been dereferenced then delete the instance
ip.

Change-Id: I9dfa62207aedc1c0a94ac7c278db5d768da11cda
Closes-Bug: #1543369