schema-transformer/config_db: Fix obj name bug
In schema-transformer/config_db.py, the method locate_route_target() of
the class RoutingInstanceST has an incorrect object reference to an
undefined object, 'rinst_obj'. It is obvious from the code that it is a
typo and should be 'self.obj' since the Routing Instance object is kept
there, and that line 1953, 1947, 1962 use this reference for the same
function call to add_route_target().

This change corrects this mistake.

We had not hit the code path earlier, but after an upgrade to Liberty,
this code path was for some reason excercised.
Liberty or not, the code path will crash any interpretation attempt by Python.

Change-Id: I336523e3b730999f3b5c0c24aca95e3d083d459c
Closes-Bug: #1543654
Signed-off-by: Martin Millnert <martin@millnert.se>