For ECMP upate from CN crash seen @CompositeNHKey::ExpandLocalCompositeNH
Problem:
When CN sends route with ECMP NH, agent creates a DB request and enqueues.
In this request there is a component NH with all local interfaces. This list is
made using the mpls label sent by CN and then querying that label to get NH
list which in trun is added in request.
Now it may happen that by the time request is processed MPLS label has been
re-used for some other non-ecmp case. During processing of this request assert
is added to check in ExpandLocalCompositeNH, if NH of MPLS label passed is
always Composite. However because of release and re-use of label as mentioned
above is possible. Hence this request should no be catered and instead wait for
another update from CN as label release will trigger a route update.

Solution:
Validate if key of NH in request is same as of MPLS label NH. If its not then
skip the request.

Closes-bug: 1458194

Conflicts:
	src/vnsw/agent/controller/controller_route_path.cc

Change-Id: I9216d52b379d80ad586d9e4277d98884f0651848