Add support for new argument to vrouter-port-control
The vif.py invokes vrouter-port-control script with an additional argument. This new
argument is valid only for mainline, but since vif.py is present in a repo which is common
for R2.20, R2.21.x and R.22.x, vrouter-port-control script in branches R2.20, R2.21.x and
R.22.x should be updated to accept and ignore this argument.

Change-Id: I6c839e9a797b4992f7551216ca1d22a7a77cde9e
Closes-Bug: #1543936