Removing SIGCHILD event for Snat/Lbaas
When Agent starts vrouter_netns.py script to setup the Namespace, it
currently tracks two events : SIGCHILD and Pipe. Pipe is used to read
the errors from the scripts. Signal is used to identify whether the
child process exited or not. It can so happen that Signal can be
recceived first and error data from pipe can arrive little later. As
part of Signal InstanceTask objected is deleted, and in error datta from
pipe if we access this InstanceTask object we end up in having a crash.

As a fix the signal event has been removed and only pipe events are read
to identify whether the child process exited or not. When error returned
on pipe asio scoket is ENOENT, the event is treated as child process
exit. Due to this, the InstanceTask object is deleted only from single
path.

Change-Id: I19c29103c54c1429216ad0022ce104089fa1f6f4
closes-bug: #1542159