SRIOV Uves
Agent is not generating any UVE's for SRIOV ports as the interface are
not added to OperDB due to nova plug information missing. With these
changes, a new VMInterfaceType and DeviceType is added as SRIOV to add
the interface to oper without nova information. Also L2, L3 routes
generatio, mpls label generation, health check on this have been
disabled as Agent does not do any of these for SRIOV interfaces.

Change-Id: I4677e31f31e176bdaa2a044d81aeb393107be55d
partial-bug: #1544198