Recreate link to route aggregate and policy in RI-ST constrtor
When routing instance is read during init time, we were not recreating any links
to the constructor. When such a routing instance is deleted, it will have those
back refs and we won't be able to delete it. If we recreate the links, we will
know to delete those links before deleting the routing instance.

Change-Id: I3938a2dfed27b084f2aa9c031ae39ea39400d33f
Closes-Bug: 1544826