Move active_flows from VrouterStatsAgent to VrouterStatsAgent.flow_rate
This is required by UI to chart a graph based on active flow count. Moving ensures that
series of flow counts information is available for UI to query because new location is under
flow_rate which has tags annotation. ‘active_flows’ should be interpreted as ‘dial’ type as
aggregate value is sent always.

Change-Id: Ic89e8cd2691a14529b1455b9533930ed0437fb84
Closes-Bug: #1545654