Correct the metadata string of fip-pool to vn
The metadata mentioned in xsd for the link between floating-ip-pool to
VN is "virtual-network-floating-ip-pool" and Agent is adding a rule
in dependency manager for VMI as "floating-ip-pool-virtual-network".
This is resulting in no InterestedEvent in dependency tracker if the
link added between VN-FIP is the last link in that VMI.

As a fix, metadata is corrected in Agent.

Change-Id: I484099d8c843493cfcdb618a3165819f9eab5d08
closes-bug: #1544788