Trigger route aggregation on origin vn match
The more specific or contributing route to a route aggregate should have same origin
vn as what is calculated for aggregated route.

In case of route aggregation on service chain, the origin vn for aggregated route
is dest-vn of the service chain config. So contributing routes are considered
only if it re-originated/service chain route.

This fix ensure that route aggregation is triggered only with service chain routes

Added UT to validate service chain and route aggregation on routing instance object

Change-Id: Ifa210749def66b033400777e5d9d0c09ade65ebe
Closes-bug: #1547975