BGP session in BGPAAS fails sometimes.
Problem:
Say there is a forward flow (F1) with Nat'd reverse flow (F2).
For some reason session from VM to BGP server does not gets created and
client(VM) keeps sending sync messages. When the source port changes in sync
message after couple of re-tries, a new flow (F3) is to be created. The reverse
flow for F3 is still F2, as key is same for F2. Because of this update,F3 and F2
should get linked but it does not happen.
This happens because agent while creating F3 tries to add F2 again instead of
issuing a change. Due to this flow index manager tries to evict flow F2(because
its add for F2 and index is present).This F2 does not get deleted as F3 is
holding reference and in turn F2 is dangling in deleted state. Update never goes
to vrouter.

Solution:
Send update for F2 when link is to be updated between forward and reverse flow.

Closes-bug: 1533924
Change-Id: I84ed4555cdd91902e85519546a7b540a814c7259