Notify replicated routes
Scenario:

Consider a route with community c1 and c2 and routing instance with policy P1
Policy P1: { from community c1; then local-pref 9999 }
Due to the routing policy the route has local-pref set 9999. The route in
bgp.l3vpn.0 table has two paths (one replicated and other Bgp path) with
local-pref set to 9999.

Now attach routing instance to policy P2
Policy P2: { from community c2; then local-pref 8888 }

Since the route has community c1 and c2, it matches both policy and the end
result is path has local-pref set to 8888.

When the new route is replicated to bgp.l3vpn.0 table, new replicated route
will have lower local-pref(8888) compared to bgp-path from other control-node(9999).
So the replicated path will not be selected and is not inserted on top of the
path list. In this case if the replicated route in bgp.l3vpn.0 table is not
notified, then export code will not send the updated path with new local-pref

Fix:
Notify replicated even when inserted secondary path is not BestPath on the route.

Added test to verify that after the policy update on routing instance, the route
has local-pref as per the policy action update(i.e. in the above example 8888)

Change-Id: I762e001ba97d4739d12daf580109f2b042878ddf
Closes-bug: #1548115