Handle KeyError if aap json data is missing mac field
Currently UI and neutron plugin db is sending aap mac as empty string "". This
scenario is handled in the current code. But if the mac_address field is missing
in the input json data or is set as "NULL", api server doesn't handle it.
In case of missing field, KeyError is thrown and in case of "NULL", the mac_address
is not defaulted to port mac.

Fix the parsing of mac_address field of aap config

Change-Id: I1c48406a3d21376cc852b605a8ce14700c17f1e4
Closes-bug: #1549339