Ecmp path should publish community
In case of service scaling with service VMs launched in same compute node, agent
publishes ecmp path. This path should contain the community set to individual VM
path. Logic was missing to copy the community attribute.

Fix the AllocateEcmpPath & ModifyEcmpPath to copy the individual path's community

TODO: This only fixes the case of allocation of ecmp path.
InetUnicastRouteEntry::EcmpAddPath  is missing logic to update the attributes of
ecmp path when individual path is updated

Change-Id: I0ac37689e380bacef40b16a714ed11bba3565658
Partial-Bug: 1549541