Reinitialize routing instance refs in routing policy on start up
When schema transformer is restarted, we are not reinitializing link
betweeen routing instance and routing policy objects. As a result, if
we later try to delete the routing instance, it fails because of existing
refs to routing policy. With this fix, we will delete the ref first
before deleting the routing instance.

Change-Id: Iceb0ff46630076fcd0480808605f56a22f62036e
Closes-Bug: 1544826
(cherry picked from commit c094b23914bfed5c3a33920c25fd93c966173418)