Converting the VMI bindings to Json format
Right now the VMI keyvalue pairs like port profile and vif details are
dict values rather strings. These are written to cassandra DB as is
without converting to Json format. As part of PotShow APi, the data is
returned as is and because of which, UI fails to display these values.

As a fix, the data is written to DB in Json format and converted back
while after retrieving from DB.

closes-bug: #1539977

Change-Id: I13846c082b42e446ee337b2b1b2df1519d4aaf08