Use service list also as key for service chains
Currently, ServiceChain class does not use service list as a key.
As a result, if there are two policy rules with the same match condition,
but different service lists, we will keep on deleting one service chain
and creating another one. With this change, the two rules will result in
two separate service chains and they should be indpendently created and
deleted and should not interfere with each other.

Change-Id: If4009330d32ab885fb64cfe98adf551e00662c53
Closes-Bug: 1508044