Send TOR mastership info in Prouter UVE.
Defined a field named connected_agent_list in Prouter UVE. Master agent
will populate this field with its name as the only element in the list. Backup
agents will send empty list. This list can be aggregated by analytics module
and if the size of the list is 0 or more than 1, alarm can be raised.

Change-Id: I01a995682eb0e23cd52ee6724fad4249586773ed
Partial-Bug: #1550888