Handle redis restart in alarm-gen
Presently, alarm-gen doesn't identify/handle redis-server restart case.
On redis-server restart, the aggregated UVEs are flushed.
alarm-gen doesn't identify the redis-server restart explicitly. It
handles the restart case, only if there is any exception on write due to
inconsistency in the local aggregated UVE cache and the redis aggregated
UVE db. Also, upon receiving UVE notification from kafka, alarm-gen does
UVE aggregation and compares it with its local aggregated UVE cache
before updating it in redis. Therefore, if there is no change in the
UVE after redis restart, it would not be updated in redis.

This patch adds support for handling redis-server restart in alarm-gen.
When alarm-gen connects to the local redis-server, it sets the key
"ALARMGEN" to True. In run_uve_processing(), on every iteration, if the
redis instance is not None, it checks if the "ALARMGEN" key exists to
identify the redis restart case and subsequently invoke
reconnect_agg_uve() to cleanup the local aggregated UVE cache and reset
the partition handlers.

Change-Id: Ib80d4662003a69d948d848f0334055bdbea435d0
Closes-Bug: #1554620