Re-initialize FlowStatsCollector pointer of FlowEntry.
During flow entry allocation from free_list, the fsc_ pointer of
FlowEntry was not being initialized. Because of state fsc_ pointer
agent was crashing.

Also add changes to return FlowStatsCollector which is not marked
for delete, during FlowStatsCollector pointer assignment to FlowEntry.
Also ensure that forward and reverse flows go to same FlowStatsCollector.
Closes-Bug: #1552826

Change-Id: I8950f3799944d67c3b8af9c7bfdb7a3ee1f01d50