Enqueue flow packet post packet parse to ensure distribution across threads
Problem-1:
When packet is enqueued to flow module without parsing the packet. As a
result, all request are enqueued to thread-0.

Fix:
Enqueue packet to PktHandler::ProcessPacket so that packet gets parsed
and then enqueued to flow module

Going further, need to ensure packet is parsed in ASIO context and
enqueued to right flow-thread from ASIO context

Problem-2:
Flows that need port-translation are always processed in thread-0. If
the flow is initially enqueued for other threads, the pkt-handler module
will enqueues flow to thread-0 with event REENTRANT.

Linklocal flows where port translation happens are not being treated as
REENTRANT flows due to bug in identifying REENTRANT flows.

Problem-3:
Flows enqueued with command REENTRANT do not have buffer. As a result,
there is a crash in PktHandler::Run() trying to access content of packet
buffer and parse the packet
Fix:
Move the packet parse part of code to VROUTER_FLOW_MSG events only

Change-Id: I78324c30b0753ee468703e5185cdf11d9db1a32f
Partial-Fix: #1522973