Svc_monitor scheduling enhancements
Currently every launch of a netns VM causes fetch of vrouter UVEs.
This is very inefficient and the plan is to maintain a local cache for
the same. Also need to add debouncing logic for marking the vrouter
as down as the collectors might return inconsistent data when there
is a flux in the system.

Also added fix to pick analytics api client from discovery and not
from configuration as before.

Also added a fix in discovery_client to skip encoding ints to
utf-8.

Adding the missing SConscript file to include the new test file.
Closes-Bug: 1536352

Change-Id: I27a9dc6721e8c65234ffa5612cf70483d03141df