Do not create right service RI for nat instances
If a service instance is in-network-nat mode, the traffic on the right side
is always routed in the primary RI. The service RI is not used for anything.
However, since all routes from primary RIs will still be copied into it.
With this commit, we won't create the right RI for such instances.

(cherry picked from commit 0db0a7186e59b2a2115200a61066ff32a1c92322)

Do not set VRF assign rules for right interfaces of nat instances

We are planning not to link service RI with the primary RI of the
right networks of in-network-nat instances. We should also not
set VRF assign rules.

Partial-Bug: 1554175
Closes-Bug: 1562200
(cherry picked from commit 0354c2bd8177d6e15f9d6e7621e58dfebf656cd3)

Change-Id: I3c043fcf8a9b585acac8ea8bcb449ea5c91879d6