Check if client exists before adding an entry in the pending list.
Consider the following sequence:
1. VM-subscribe is received from client C1 and queued for execution via
ChannelEventProcTask.
2. XMPP-NotReady is received on channel to C1 and queued for execution via
ChannelEventProcTask.
3. Due to #1, IFMapXmppChannel::ProcessVmSubscribe() runs and queues the
VmSubscribe to task IFMapServer::IFMapVmSubscribe.
4. #2 causes ProcessChannelNotReady() to call ProcessClientWork() which cleans
up the client. The client-cleanup code checks the pending list but
IFMapServer::IFMapVmSubscribe from #3 has not executed and has not added it to
the pending list.
5. IFMapServer::IFMapVmSubscribe finally runs and adds the VM to the pending
list, even though the client has been deleted since it does not check if the
client exists or not.
6. Config finally adds the VM node. The UUID-mapper receives this VM-node-add
and finds the VM in the pending list. It expects the client to exist but the
client was deleted in #4. Hence, it asserts.

Fix is to process the VM subscribe only if the client exists. This makes sure
that we insert the VM in the pending list only if the client is alive.
Added a test that, without the fix, asserted in the same place as the core.
Also, added a test for the case where we receive config, then client gets
deleted, then we process the vm-subscribe. Without the fix, we would add VR and
VM nodes with origin XMPP, even though the client is gone.

Closes-Bug: #1562206

Change-Id: I934d0c917fa1a1955f5e1ffddb4ee14160ab5ad2
(cherry picked from commit 163ab4328d345e6c7c6656d2c97f26ba2e8adb5e)