Defer bgp peer deletion by taking pre-emption into account
BGP Peer deletion should be deferred until all close paths are complete.
This is controlled by different booleans such as defer_close_ in bgp_peer.cc
and some in bgp_peer_close.cc. But, there is a time lag between when
defer_close_ is reset and when bgp_peer_close is notified to proceed
(via task trigger_). Hence, we must also check if the task is triggerd to
deduce whether the close is still in progress.

Change-Id: Ied61f970b509241a80728a8b5f0510711b86eb28
Closes-Bug: #1563673