Create vrouter state cache from polling uve
Currently every launch of a netns VM causes fetch of vrouter UVEs.
This is inefficient and the plan is to maintain a local cache for
the same. Also need to add debouncing logic for marking the vrouter
as down as the collectors might return inconsistent data when
there is flux in the system.

Also added fix to pick analytics api client from discovery and not
from configuration as before.

Change-Id: I5c158d152355eaaee3a7b582ba41841f6fe49573
Closes-Bug: #1536821