DiscoveryClient stops resusbcribing even after ttl period.
On no response from discovery server and no error, subscribe timer is not restarted as
the subscribe_cb_called_ flag was not reset.  Setting the flag appropriately on every
subscribe request, so subscribe timer can be restarted during the case above.

HeartBeat will not be effected as heartbeat timer is periodic and is
started on succesful publish.

Remove unnecessary curl timeout traces, use stats counters.

Change-Id: Iddd5a3b4f09a6c9c6fbce14878624ada622b3174
Closes-Bug:1563002