Temporary workaround for FIP + SNAT scaling problem
Ignore connection links for non-default routing instances of virtual
networks with router-external set. This stops unnecessary import of
routes into non-default instances of external networks and subsequent
download to all vrouters that have FIP or SNAT. The instances are not
used in the forwarding path anyway.

Enable this behavior under control of a new optimize_snat option in
contrail-control.conf.

Proper fix is to not create service chains for SNAT and to not create
these service routing instances for service chains where the last SI
is a NAT (Launchpad bug 1562200).

Change-Id: Ie64bae9a7b2284b36e0b26563da4677eaa7f9157
Partial-Bug: 1554175
Related-Bug: 1562200