Core at AgentDnsXmppChannel::WriteReadyCb
Application should not free data sent over TCP socket as the library takes care
of memory management of the data.

On a socket write block xmpp provides the facility to register a callback so application
can handle the case. In case of agent, the inherent TCP library buffering should be
sufficient and no action needed from the application as there is only one writer(DNS)
on the socket.

Change-Id: I73993d58ed84cdce97217038c3da17ac3f5142d2
Closes-Bug: 1561843