Removed support for href linking - when aggregation is done in alarmgen, href linking will never work
Added support for union aggregation for maps.
This will be needed for DerivedStats.

It will allow this kind of definition:

struct SandeshUVEData {
    1: string name (key="ObjectGeneratorInfo")
    2: optional bool deleted
    3: optional map<string,i32> mapx
    4: optional map<string,derived_stats_results.NullResult> null_mapx (stats="mapx:")
}

Having the raw and derived stat be different optional attributes in the UVE struct
makes it more efficient to send the raw without the derived and the derived without the raw.

The other methods of having raw and derived on level deeper that the UVE Struct will be removed later:

struct TestStat {
    1: string sname (aggtype="listkey")
    2: i32 s
    3: derived_stats_results.NullResult null_s (stats="s:")
}

struct SandeshUVEData {
    1: string name (key="ObjectGeneratorInfo")
    2: optional bool deleted
    5: optional TestStat ts
    6: optional list<TestStat> tsl
}

Change-Id: I55535a69968f060aa2b03e8d9fc2347203e275b0
Closes-Bug: 1428271