Task exclusions to apply discovery response and trigger deletion of Managed XmppClient object.
DiscoveryClient Response which triggers Managed Deletion of XmppClient object and dependents execute in the
context of "http client" task and the ManagedDelete of the XmppClient object happens in the context
of "bgp::config" task.

As the two tasks are not exclusive, trigger of ManagedDelete will enqueue the object and task to dequeue and
delete the object may excute right away while the "http client" task is running and executing code
that triggers the MayDelete of the object.

To resolve the issue, DiscoveryResponse for DNS will also be enqueued as part of Agent::ControllerXmpp task which
is exclusive to bgp::config task.

Change-Id: I8093dc7562837ee2d0a4f959dfbf3ecb93886ed6
Closes-Bug:1560604
Closes-Bug:1562964
Closes-Bug:1563618