Limit retries on pycassa connection pool
Currently, pycassa connection pool has unlimited retries. As a result, if
cassandra server goes down, this condition is not detected and the function
hangs forever. With this change, we limit the retries to 30. After that, an
exception will be thrown, which can be caught to return appropriate error to the
caller.

Change-Id: I49050e4430996b8d6686a28e17da67cedd7a58ae
Closes-Bug: 1565681
(cherry picked from commit cb0d0732c1b76ffa73c190a9c71611f0686946f3)