BGPaaSv2: schema transformer changes.
Populate gateway address in the bgpaas-client bgp_router to make the bgp
session bidirectional in v2 of bgp-as-a-service feature.

Setting ipv4-mapped ipv6 gateway address in case of
bgpaas-ipv4-mapped-ipv6-nexthop enabled bgp-as-a-service object.

Gateway addresses are not set when bgpaas-suppress-route-advertisement
property is set in bgp-as-a-service object to support v1 behaviour.

Partial-Bug: 1563075

Change-Id: Iec58cff7dd9fad778e9f4a45c42a54d894a9b483