Introduce token-based scheduling for flow-process queues
VRouter is consumer of add/update/delete events of flows. The channel
between agent and VRouter can be a bottleneck when large number of
events are generated (due to ageing, config-change etc..). When large
number of events are generated, it can result in starving of other
operations.

A token-based scheme is introduced with this commit. Add/Update/Delete
happen thru different work-queues. The work-queues can be scheduled to
run only if they have token for corresponding operation. The tokens are
consumed when add/update/delete operation are enqueued to VRouter and
they are freed when VRouter operation compeletes. This acts a feedback
and throttling mechanism from VRouter to Agent.

Change-Id: Ief5d7873f3431d7724fc32a2510bdd7ab8f20123
Partial-Bug: #1567802