Move active_flows from VrouterStatsAgent to VrouterStatsAgent.flow_rate
This is required by UI to chart a graph based on active flow count.
Moving ensures that series of flow counts information is available for UI to
query because new location is under flow_rate which has tags annotation.
active_flows should be interpreted as dial type as aggregate value is sent
always.
                                                                                                                                                             Closes-Bug: #1545654
Change-Id: Ib5adadc8c9073bc4014d6339ff8dce4a3b776d6b